# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat/{locale}/conversations:
    post:
      operationId: createChatConversation
      tags: [chat]
      summary: Start a chat conversation
      description: Creates a new visitor conversation for the requested locale.
      parameters:
        - $ref: "#/components/parameters/LocalePath"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationInput"
      responses:
        "200":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatConversation"
        "400":
          $ref: "#/components/responses/BadRequest"
  /chat/conversations/{id}:
    get:
      operationId: getChatConversation
      tags: [chat]
      summary: Get a conversation with its messages
      description: Returns a conversation and its ordered messages for session restore.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatConversationDetail"
        "404":
          $ref: "#/components/responses/NotFound"
  /chat/conversations/{id}/messages:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message and stream the grounded reply
      description: >-
        Streams the assistant reply as Server-Sent Events. Each event is a JSON
        object with either a `content` delta or `done: true`. Consumed on the
        client with fetch + ReadableStream, not a generated hook.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageInput"
      responses:
        "200":
          description: SSE stream of reply deltas
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /chat/conversations/{id}/lead:
    post:
      operationId: captureLead
      tags: [chat]
      summary: Capture a contact lead within a conversation
      description: Stores a contact captured inline during a conversation.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaptureLeadInput"
      responses:
        "200":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  parameters:
    LocalePath:
      name: locale
      in: path
      required: true
      description: Language for the returned content.
      schema:
        $ref: "#/components/schemas/Locale"
...
  schemas:
...
    Locale:
      type: string
      enum: [en, nl]
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
      required:
        - role
        - content
        - createdAt
    ChatConversation:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: string
        locale:
          $ref: "#/components/schemas/Locale"
      required:
        - id
        - sessionId
        - locale
    ChatConversationDetail:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: string
        locale:
          $ref: "#/components/schemas/Locale"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
      required:
        - id
        - sessionId
        - locale
        - messages
    CreateConversationInput:
      type: object
      properties:
        sessionId:
          type: string
    SendMessageInput:
      type: object
      properties:
        content:
          type: string
          minLength: 1
      required:
        - content
    CaptureLeadInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          minLength: 3
        company:
          type: string
        message:
          type: string
      required:
        - name
        - email
...
    Lead:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: ["integer", "null"]
        name:
          type: string
        email:
          type: string
        company:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        locale:
          $ref: "#/components/schemas/Locale"
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - locale
        - status
        - createdAt
...